add index.html style.css or script.js and image folder
This website has been created for proposing to your loved one (wife or girlfriend) through sections like Our Story: Version History, Reasons I Love You: The API of My Heart, and Our Memory Cache. Not only that, it is specially made for software developers and computer engineers. If your partner is in this field, you can use it.